fix: vendor-in code from stringcase
Closes #236.

The python-stringcase dependency this package has is very badly
maintained:
  * There are no releases.
  * There are no git tags.
  * The last release on PyPi was 3 years ago, although internal version
    numbers have been bumped since then.
  * The project seems somewhat dormant, if not abandonned.

Having it as a dependency is thus a liability and makes this library
hard to package in Linux distributions, as packaging badly maintained
libraries is no fun.

Since the code used is actually very simple, it makes sense to vendor it
in and not depend on it anymore.

Note that I am not the author of the majority of the code in this patch
and mainly did some glue work. Credit and copyright attributions (under
the MIT) should be attributed to the original author, Taka Okunishi
<okunishinishi@gmail.com>. I tried to make that clear by adding a
copyright header to the new stringcase.py file.

The original code can be found at:
https://raw.githubusercontent.com/okunishinishi/python-stringcase/master/stringcase.py